+++ crates/herkos/src/ir/types.rs
@@ -176,14 +176,14 @@ pub enum IrInstr {
     /// Call direct function (local, not imported)
     Call {
         dest: Option<VarId>, // None for void functions
-        func_idx: u32,       // Local function index (imports are handled separately)
+        func_idx: usize,     // Local function index (imports are handled separately)
         args: Vec<VarId>,
     },
 
     /// Call imported function from host
     CallImport {
         dest: Option<VarId>, // None for void functions
-        import_idx: u32,     // Index into the imports list
+        import_idx: usize,   // Index into the imports list
         module_name: String, // Import module name (e.g., "env")
         func_name: String,   // Import field name (e.g., "log")
         args: Vec<VarId>,
@@ -192,7 +192,7 @@ pub enum IrInstr {
     /// Call indirect (via table)
     CallIndirect {
         dest: Option<VarId>,
-        type_idx: u32,
+        type_idx: usize,
         table_idx: VarId,
         args: Vec<VarId>,
     },
@@ -202,10 +202,10 @@ pub enum IrInstr {
     Assign { dest: VarId, src: VarId },
 
     /// Read a global variable (dest = globals.g{index} or const G{index})
-    GlobalGet { dest: VarId, index: u32 },
+    GlobalGet { dest: VarId, index: usize },
 
     /// Write a mutable global variable (globals.g{index} = value)
-    GlobalSet { index: u32, value: VarId },
+    GlobalSet { index: usize, value: VarId },
 
     /// Query current memory size in pages (dest = memory.size())
     MemorySize { dest: VarId },
@@ -666,7 +666,7 @@ pub struct FuncExport {
     /// The exported name (becomes a Rust method name).
     pub name: String,
     /// Index into the function index space.
-    pub func_index: u32,
+    pub func_index: usize,
 }
 
 /// Signature of a function.
@@ -677,7 +677,7 @@ pub struct FuncSignature {
     /// Return type (None for void).
     pub return_type: Option<WasmType>,
     /// Index into the Wasm type section (needed for call_indirect dispatch).
-    pub type_idx: u32,
+    pub type_idx: usize,
     /// Whether this function calls imported functions (needs host parameter).
     pub needs_host: bool,
 }
@@ -686,9 +686,9 @@ pub struct FuncSignature {
 #[derive(Debug, Clone)]
 pub struct ElementSegmentDef {
     /// Starting offset in the table.
-    pub offset: u32,
+    pub offset: usize,
     /// Function indices to place into the table starting at `offset`.
-    pub func_indices: Vec<u32>,
+    pub func_indices: Vec<usize>,
 }
 
 /// An imported function for trait generation.
@@ -727,13 +727,13 @@ pub struct ModuleInfo {
     /// Whether the module declares linear memory.
     pub has_memory: bool,
     /// Maximum memory pages (from Wasm memory section or default).
-    pub max_pages: u32,
+    pub max_pages: usize,
     /// Initial memory pages (from Wasm memory section).
-    pub initial_pages: u32,
+    pub initial_pages: usize,
     /// Initial table size (number of entries).
-    pub table_initial: u32,
+    pub table_initial: usize,
     /// Maximum table size (for const generic TABLE_MAX).
-    pub table_max: u32,
+    pub table_max: usize,
     /// Element segments for table initialization.
     pub element_segments: Vec<ElementSegmentDef>,
     /// Global variable definitions (mutable + immutable).
@@ -749,7 +749,7 @@ pub struct ModuleInfo {
     /// Canonical type index mapping: maps each Wasm type index to the
     /// smallest index with the same structural signature.
     /// Used for spec-compliant structural type equivalence in call_indirect.
-    pub canonical_type: Vec<u32>,
+    pub canonical_type: Vec<usize>,
     /// Imported functions for trait generation.
     /// The number of imported functions is `func_imports.len()`.
     pub func_imports: Vec<FuncImport>,
@@ -765,8 +765,8 @@ impl ModuleInfo {
     /// Get the number of imported functions.
     ///
     /// This is derived from `func_imports.len()` rather than storing it separately.
-    pub fn num_imported_functions(&self) -> u32 {
-        self.func_imports.len() as u32
+    pub fn num_imported_functions(&self) -> usize {
+        self.func_imports.len()
     }
 
     /// Whether the module needs a wrapper struct (Module/LibraryModule).
